+++ chartlets.js/packages/lib/src/plugins/vega/index.ts
@@ -1,0 +1,6 @@
+import type { Plugin } from "@/index";
+import { VegaChart } from "./VegaChart";
+
+export default function vega(): Plugin {
+  return { components: [["VegaChart", VegaChart]] };
+}